+++ Assets/Master/Script.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d05405299586644195442c4fac6f75d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
